requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.26.0
asynctest==0.13.0
attrs==22.1.0
charset-normalizer==2.1.0
databases==0.6.1
flake8==5.0.4
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.2.0
marshmallow==3.17.0
mccabe==0.7.0
multidict==6.0.2
packaging==21.3
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
SQLAlchemy==1.4.40
typing_extensions==4.3.0
yarl==1.8.1
zipp==3.8.1
